Add social ratings & geohash support to Home
Wire review averages and user geohash into home ranking and UI. Added a ReviewDao query to compute average ratings by a list of authors and changed PubkeyCount.count from Int to Float to hold averages. HomeViewModel now accepts ReviewDao, exposes userGeohash state, fetches the current profile geohash, builds a socialRatingsMap from reviews, and passes currentGeohash/socialRatingsMap into the MerchantRankingEngine; sorting for "Nearest" and "Top Rated" was updated accordingly. HomeScreen was updated to provide the reviewDao to the VM, observe userGeohash, and show explanatory messages for "Top Rated" and "Nearest" sorts. Note: PubkeyCount type change may require adjustments/migration where it is consumed.

+++ app/src/main/java/com/ely/kian/data/local/entities/Profile.kt
@@ -40,5 +40,5 @@ data class UserFollow(
 
 data class PubkeyCount(
     val pubkey: String,
-    val count: Int
+    val count: Float
 )
